+++ src/examples/fileserver_example.c
@@ -29,8 +29,8 @@
 
 #define PAGE "<html><head><title>File not found</title></head><body>File not found</body></html>"
 
-static int
-file_reader (void *cls, uint64_t pos, char *buf, int max)
+static ssize_t
+file_reader (void *cls, uint64_t pos, char *buf, size_t max)
 {
   FILE *file = cls;
